+++ apps/fluent-tester/src/TestComponents/RadioGroupExperimental/RadioGroupExperimentalTest.tsx
@@ -3,6 +3,7 @@ import { RADIO_GROUP_EXPERIMENTAL_TESTPAGE } from './consts';
 import { DefaultRadioGroup } from './DefaultRadioGroup';
 import { RequiredRadioGroup } from './RequiredRadioGroup';
 import { DisabledRadioGroup } from './DisabledRadioGroup';
+import { SubtextRadioGroup } from './SubtextRadioGroup';
 import { HorizontalRadioGroup } from './HorizontalRadioGroup';
 import { Test, TestSection, PlatformStatus } from '../Test';
 
@@ -23,10 +24,15 @@ const radioGroupExperimentalSections: TestSection[] = [
     component: RequiredRadioGroup,
   },
   {
+    name: 'RadioGroup with Label Subtext',
+    testID: RADIO_GROUP_EXPERIMENTAL_TESTPAGE,
+    component: SubtextRadioGroup,
+  },
+  {
     name: 'Other Layouts',
     testID: RADIO_GROUP_EXPERIMENTAL_TESTPAGE,
     component: HorizontalRadioGroup,
-  },
+},
 ];
 
 export const RadioGroupExperimentalTest: React.FunctionComponent = () => {