+++ packages/experimental/RadioGroup/src/Radio/Radio.styling.ts
@@ -1,7 +1,8 @@
 import { radioName, RadioTokens, RadioSlotProps, RadioProps } from './Radio.types';
-import { UseStylingOptions, buildProps } from '@fluentui-react-native/framework';
+import { Theme, UseStylingOptions, buildProps } from '@fluentui-react-native/framework';
 import { globalTokens } from '@fluentui-react-native/theme-tokens';
 import { defaultRadioTokens } from './RadioTokens';
+import { fontStyles } from '@fluentui-react-native/tokens';
 
 export const radioStates: (keyof RadioTokens)[] = ['labelPositionBelow', 'focused', 'hovered', 'pressed', 'selected', 'disabled'];
 
@@ -13,7 +14,7 @@ export const stylingSettings: UseStylingOptions<RadioProps, RadioSlotProps, Radi
       (tokens: RadioTokens) => ({
         style: {
           display: 'flex',
-          alignItems: 'center',
+          alignItems: 'flex-start',
           flexDirection: tokens.flexDirection,
           minHeight: 20,
           marginTop: 0,
@@ -57,15 +58,35 @@ export const stylingSettings: UseStylingOptions<RadioProps, RadioSlotProps, Radi
       }),
       ['radioInnerCircleSize', 'radioVisibility', 'radioFill'],
     ),
+    labelContent: {
+      style: {
+        alignSelf: 'center',
+        alignItems: 'flex-start',
+        flexDirection: 'column',
+      },
+    },
     label: buildProps(
-      (tokens: RadioTokens) => ({
+      (tokens: RadioTokens, theme: Theme) => ({
         variant: tokens.variant,
         style: {
           marginTop: tokens.marginTop,
           color: tokens.color,
+          ...fontStyles.from(tokens, theme),
         },
       }),
-      ['color', 'marginTop'],
+      ['variant', 'color', ...fontStyles.keys],
+    ),
+    subtext: buildProps(
+      (tokens: RadioTokens) => ({
+        variant: tokens.subtextVariant,
+        style: {
+          marginTop: tokens.subtextMarginTop,
+          marginRight: tokens.subtextMarginRight,
+          marginBottom: tokens.subtextMarginBottom,
+          color: tokens.color,
+        },
+      }),
+      ['subtextMarginTop', 'subtextMarginRight', 'subtextMarginBottom', 'color', 'subtextVariant'],
     ),
   },
 };