+++ packages/experimental/RadioGroup/src/Radio/Radio.types.ts
@@ -4,6 +4,7 @@ import { FontTokens, IForegroundColorTokens, IBackgroundColorTokens, IBorderToke
 import { IPressableProps } from '@fluentui-react-native/pressable';
 import { IFocusable, IPressableHooks } from '@fluentui-react-native/interactive-hooks';
 import { ColorValue, ViewStyle } from 'react-native';
+import { Variant } from '@fluentui-react-native/framework';
 
 export const radioName = 'Radio';
 
@@ -42,11 +43,33 @@ export interface RadioTokens extends FontTokens, IColorTokens, IForegroundColorT
    * The flex direction of the root
    */
   flexDirection?: ViewStyle['flexDirection'];
+  
+   * Variant of label subtext.
+   *
+   * Should only by used if subtext prop is provided.
+   */
+  subtextVariant?: keyof Variant;
 
   /**
-   * Top margin
+   * Padding between label and label subtext.
+   *
+   * Should only by used if subtext prop is provided.
    */
-  marginTop?: ViewStyle['marginTop'];
+  subtextMarginTop?: ViewStyle['marginTop'];
+
+  /**
+   * Padding between label subtext and focus ring.
+   *
+   * Should only by used if subtext prop is provided.
+   */
+  subtextMarginRight?: ViewStyle['marginRight'];
+
+  /**
+   * Padding between label subtext and focus ring.
+   *
+   * Should only by used if subtext prop is provided.
+   */
+  subtextMarginBottom?: ViewStyle['marginBottom'];
 
   /**
    * States that can be applied to a Radio
@@ -64,6 +87,11 @@ export interface RadioProps extends IPressableProps {
    * The text string for the option
    */
   label: string;
+
+  /**
+   * Label subtext for the option
+   */
+  subtext?: string;
 
   /**
    * A unique key-identifier for each option
@@ -103,7 +131,9 @@ export interface RadioSlotProps {
   root: IViewProps;
   button: IViewProps;
   innerCircle: IViewProps;
+  labelContent: IViewProps;
   label: TextProps;
+  subtext?: TextProps;
 }
 
 export interface RadioType {