+++ packages/experimental/RadioGroup/src/RadioGroup/RadioGroup.tsx
@@ -60,7 +60,7 @@ export const RadioGroup = compose<RadioGroupType>({
       if (children) {
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore - TODO, fix typing error
-        contextValue.buttonKeys = React.Children.map(children, (child: React.ReactChild) => {
+        contextValue.values = React.Children.map(children, (child: React.ReactChild) => {
           if (React.isValidElement(child)) {
             return child.props.value;
           }