+++ packages/components/Persona/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Change Log - @fluentui-react-native/persona
 
-This log was last generated on Thu, 06 Oct 2022 17:24:49 GMT and should not be manually modified.
+This log was last generated on Tue, 11 Oct 2022 22:41:44 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.10
+
+Tue, 11 Oct 2022 22:41:44 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.12.10
+- Bump @fluentui-react-native/framework to v0.8.9
+- Bump @fluentui-react-native/persona-coin to v0.11.10
+- Bump @fluentui-react-native/tokens to v0.17.3
+- Bump @uifabricshared/foundation-tokens to v0.12.6
 
 ## 0.13.9