+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,31 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Mon, 10 Oct 2022 23:26:12 GMT and should not be manually modified.
+This log was last generated on Tue, 11 Oct 2022 22:41:44 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.32.32
+
+Tue, 11 Oct 2022 22:41:44 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.25.7
+- Bump @fluentui-react-native/callout to v0.21.20
+- Bump @fluentui-react-native/checkbox to v0.16.23
+- Bump @fluentui-react-native/contextual-menu to v0.18.22
+- Bump @fluentui-react-native/focus-trap-zone to v0.9.19
+- Bump @fluentui-react-native/focus-zone to v0.10.20
+- Bump @fluentui-react-native/interactive-hooks to v0.18.12
+- Bump @fluentui-react-native/link to v0.16.18
+- Bump @fluentui-react-native/menu-button to v0.8.41
+- Bump @fluentui-react-native/persona to v0.13.10
+- Bump @fluentui-react-native/persona-coin to v0.11.10
+- Bump @fluentui-react-native/pressable to v0.9.18
+- Bump @fluentui-react-native/radio-group to v0.15.22
+- Bump @fluentui-react-native/separator to v0.12.9
+- Bump @fluentui-react-native/tabs to v0.7.21
+- Bump @fluentui-react-native/text to v0.15.5
 
 ## 0.32.31